configurable fee + moved some storages
